requirements.txt
beautifulsoup4==4.6.0
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dataclasses==0.7
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
gevent==20.5.2
google-api-core==1.17.0
google-auth==1.14.3
google-cloud-core==1.3.0
google-cloud-translate==2.0.1
googleapis-common-protos==1.51.0
googletrans==2.4.0
greenlet==0.4.15
grpcio==1.29.0
idna==2.9
joblib==0.16.0
JPype1==0.7.5
konlpy==0.5.2
lxml==4.5.1
mecab-python===0.996-ko-0.9.2
nltk==3.5
numpy==1.19.2
oauthlib==3.1.0
olefile==0.46
packaging==20.4
pandas==1.0.4
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1594212964103/work
protobuf==3.12.0
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
sacremoses==0.0.43
scikit-learn==0.23.1
scipy==1.5.1
sentencepiece==0.1.91
six==1.14.0
soynlp==0.0.493
threadpoolctl==2.1.0
tokenizers==0.9.2
torch==1.5.1
torchvision==0.6.0a0+35d732a
tqdm==4.46.1
transformers==3.4.0
tweepy==3.8.0
urllib3==1.25.9
websocket==0.2.1
zope.event==4.4
zope.interface==5.1.0
